fix(entities.sql): change entity_id type from BIGINT to INTEGER (8 bytes -> 4 bytes)

+++ easyner/io/database/schemas/entities.sql
@@ -2,7 +2,7 @@
 CREATE SEQUENCE IF NOT EXISTS entity_id_seq; -- Should be here as it's required by the entities table
 
 CREATE TABLE IF NOT EXISTS entities (
-    entity_id BIGINT DEFAULT nextval('entity_id_seq'),
+    entity_id INTEGER DEFAULT nextval('entity_id_seq'),
     article_id INTEGER,
     sentence_id INTEGER,
     text VARCHAR,
